refactor(settings/speech): Creating container and constants files
